refactor: schema questions
